pyproject.toml
[tool.poetry]
name = "pydata-prefect-workshop"
version = "0.2"
description = "Packages for completing the Prefect practical workshop at PyData"
authors = ["Adam Hill <adam.ben.hill@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
prefect = { version = "*", allow-prereleases = true }
kafka-python-ng = "^2.2.2"
httpx = "^0.26.0"
dask = "^2024.8.1"
pymongo = "^4.8.0"
nltk = "^3.8.1"
ipykernel = "^6.29.5"
wordcloud = "^1.9.3"
streamlit = "^1.37.1"
python-dotenv = "^1.0.1"
plotly = "^5.23.0"
matplotlib = "^3.9.2"
pandas = "^2.2.2"
numpy = ">=1.26,<2.0"
pip = "^24.2"
spacy = "^3.7.5"
spacytextblob = "^4.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
pytest = "^8.3.2"
coverage = "^7.6.1"
pytest-cov = "^5.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
